Big Changes: - Add class Field, make him hashmap and put him in all fields on Island; -Fill the Island; -Change mechanics; -Add threads to all animals;
