+++ src/Main.java
@@ -1,6 +1,8 @@
+
+
 public class Main {
     public static void main(String[] args) {
+        new Menu();
 
-        new Menu().start();
     }
 }
