+++ src/core/animals/herbivorous/Herbivorous.java
@@ -1,0 +1,16 @@
+package core.animals.herbivorous;
+
+import core.animals.Animals;
+
+
+public abstract class Herbivorous extends Animals {
+
+    private String animalName;
+    private String animalIcon;
+    private double weightOfAnimal;
+    private int maxStepByMove;
+    private double maxKiloCanEat;
+//    private double stomachFullness;
+
+
+}
