+++ src/core/map/Field.java
@@ -1,17 +1,15 @@
-package island.map;
+package core.map;
 
-import island.IslandFormOfLife;
+import core.IslandFormOfLife;
 
 import java.util.HashMap;
 
-public class Field extends Island{
-    private HashMap<IslandFormOfLife, Integer> field = new HashMap<>();
-
+public class Field {
+    protected static HashMap<IslandFormOfLife, Integer> field = new HashMap<>();
 
     public HashMap<IslandFormOfLife, Integer> getFieldHashMap(){
         return field;
     }
-
 
     public void setFieldHashMap(HashMap<IslandFormOfLife,Integer> field){
         this.field = field;
