--- src/island/animals/herbivorous/Mouse.java
@@ -1,117 +1,0 @@
-package island.animals.herbivorous;
-
-import island.IslandFormOfLife;
-import island.Plants;
-import island.animals.Animals;
-import island.animals.predators.Bear;
-import java.util.HashMap;
-import java.util.Map;
-
-public class Mouse extends Herbivorous  {
-    private final Map<IslandFormOfLife, Integer> canEat = new HashMap<>();
-    private final String animalName = "Mouse";
-    private final String animalIcon = "\uD83D\uDC01";
-    private final double weightOfAnimal = 0.05;
-    private final int maxPopulationOnOneLocation = 500;
-    private final int maxStepByMove = 1;
-    private final double maxKiloCanEat = 0.01;
-    private final int animalLowHealth = 40;
-    private final int chanceToEatCaterpillar = 90;
-    private final int chanceToEatPlants = 100;
-    private double stomachFullness = 0.005;
-
-
-    {
-        canEat.put(new Plants(), chanceToEatPlants);
-        canEat.put(new Caterpillar(), chanceToEatCaterpillar);
-    }
-
-    public Mouse(){
-        setAnimalName(animalName);
-        setAnimalIcon(animalIcon);
-        setWeightOfAnimal(weightOfAnimal);
-        setMaxKiloCanEat(maxKiloCanEat);
-        setStomachFullness(stomachFullness);
-        setMaxStepByMove(maxStepByMove);
-    }
-    protected void getHungry(){
-        double hungry = getMaxKiloCanEat()/10;
-        setStomachFullness(getStomachFullness() - hungry);
-    }
-    public boolean eat(Animals islandFormOfLife){
-        int chanceOfEat;
-        if(stomachFullness < maxKiloCanEat) {
-            if (canEat.containsKey(islandFormOfLife)) {
-                chanceOfEat = RANDOM.nextInt(100 + 1);
-                if (chanceOfEat >= canEat.get(islandFormOfLife)) {
-                    stomachFullness += islandFormOfLife.getWeightOfAnimal()/2;
-                    islandFormOfLife.setAlive(false);
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-
-    @Override
-    public boolean eat(Plants plant) {
-        if(canEat.containsKey(plant)) {
-            super.eat(plant);
-        }
-        return true;
-    }
-
-    @Override
-    public boolean eat(IslandFormOfLife islandFormOfLife) {
-        return false;
-    }
-
-    @Override
-    public void reproduce(Animals animals) {
-        int chanceOfReproduce = RANDOM.nextInt(2);
-        if(chanceOfReproduce == 1){
-            if(!(animalsOnOneField.size() >= maxPopulationOnOneLocation)) {
-                animalsOnOneField.add(new Bear());
-            }
-        }
-    }
-
-
-    public void die(IslandFormOfLife islandFormOfLife){
-        if(weightOfAnimal == animalLowHealth){
-            animalsOnOneField.remove(islandFormOfLife);
-        }
-    }
-
-    @Override
-    public void run(){
-        setY(RANDOM.nextInt(island.getHeight()));
-        setX(RANDOM.nextInt(island.getWidth()));
-
-        island.fields[getY()][getX()].animalsOnOneField.add(this);
-        while (isAlive()) {
-            island.move(this);
-
-            for (Animals animal : island.fields[getY()][getX()].animalsOnOneField){
-                this.eat(animal);
-            }
-
-            for (Plants plant : island.fields[getY()][getX()].plantsOnOneField) {
-                this.eat(plant);
-            }
-
-            for(Animals animal : island.fields[getY()][getX()].animalsOnOneField){
-                this.reproduce(animal);
-            }
-
-            getHungry();
-            die(this);
-
-            try{
-                Thread.sleep(3000);
-            }catch (InterruptedException e){
-                throw new RuntimeException("Thread was interrupted!");
-            }
-        }
-    }
-}
